# api-put-object-streaming: Removed unnecessary `partBuf` copy to save allocations; no race detected.