Hi and thanks for the great work so far. 👋🏽 

Following our discussion in https://github.com/minio/mc/issues/3376#issuecomment-806013131 I performed extended testing, benchmarking and profiling to figure out what's best for `minio-go` and the Thanos [issue](https://github.com/thanos-io/thanos/issues/3967). 

Proving tests:
* Race [test](https://github.com/thanos-io/thanos/pull/5474/files#diff-5e00ccdab4b3bcb9f66b512babda69ce310f217029f01a669c612c995f9e1f31R29)
* [Benchmark](https://github.com/thanos-io/thanos/pull/5474/files#diff-5e00ccdab4b3bcb9f66b512babda69ce310f217029f01a669c612c995f9e1f31R83)

No matter how hard I tried, I could not reproduce ANY race the [fix](https://github.com/minio/minio-go/pull/1357) was made for. I even try to trigger retries, and all works fine for Linux `*os.File`. I tried to look for any repro tests on your side but none was provided to prove that there is indeed any race. Perhaps the race was detected on certain OS? (windows/MacOS?)

The benchmarks show significant allocation problems while copying large files (PartSize 64MB). I also optimize naive chunk preallocations.

```
benchstat v1.txt v2-upgr.txt
name       old time/op    new time/op    delta
Upload-12     300ms ± 7%     283ms ± 7%     ~     (p=0.310 n=5+5)

name       old alloc/op   new alloc/op   delta
Upload-12    69.3MB ± 0%     0.5MB ± 3%  -99.23%  (p=0.008 n=5+5)

name       old allocs/op  new allocs/op  delta
Upload-12     3.33k ± 2%     3.16k ± 1%   -5.01%  (p=0.008 n=5+5)
```

I know you reduced default PartSize to 16MB (from 128MB) to reduce default allocation spread, but this is still unacceptable and 16MB part size have other tradeoffs (latency). So in this PR I propose to revert that fix and optimize allocations. I would love to be e able to use minio-go in Thanos and related projects, but we need this path to be leaner. One of the components Thanos sidecar is usually deployed with limited mem (100-200 MB), so optimizations in this library matters.

Perhaps if you don't want to accept this PR because you are afraid of (IMO unproven) potential races, you would be happy with adding extra option to `Put` API like `ReuseReader`? 

I am happy to be proven wrong that we indeed have raced here (e.g I did not test other OS than Linux), but we can't accept 60x allocation regression for unexisting race. Let me know what do you think, thanks! 🤗 
 
This version is also proposed to be used in Thanos: https://github.com/thanos-io/thanos/pull/5474

Fixes: https://github.com/thanos-io/thanos/issues/3967
Fixes: https://github.com/thanos-io/thanos/issues/5393