@klauspost 2022-07-06
@bwplotka 

`ReadAt` is concurrency safe on `*os.File`, but the `io.ReaderAt` interface does *not* mandate that this is safe for all implementations.

Therefore assuming that it is, can lead to subtle bugs.