@bwplotka 2022-07-06
Ack.

Would you accept an explicit option to reuse reader in options then? (: