@harshavardhana 2022-07-06
> Ack.
> 
> Would you accept an explicit option to reuse reader in options then? (:

what I would suggest is we split the function instead of a generic ReaderAt make that ReaderAt by explicit *os.File - rest of them pay the penalty of being io.Reader instead since we can't reliably guarantee that concurrency is possible.