@harshavardhana 2022-07-06
The actual issue because of a "retry" - I reproduced the problem even with this PR

```
==================
WARNING: DATA RACE
Write at 0x00c0034d2a08 by goroutine 132:
  io.(*SectionReader).Seek()
      /home/harsha/sdk/go1.18.3/src/io/io.go:534 +0x165
  github.com/minio/minio-go/v7.Client.executeMethod()
      /home/harsha/go/src/github.com/minio/minio-go/api.go:564 +0x4dd
  github.com/minio/minio-go/v7.Client.uploadPart()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object-multipart.go:282 +0x533
  github.com/minio/minio-go/v7.Client.putObjectMultipartStreamFromReadAt.func2()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object-streaming.go:171 +0x651
  github.com/minio/minio-go/v7.Client.putObjectMultipartStreamFromReadAt.func3()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object-streaming.go:192 +0x47

Previous write at 0x00c0034d2a08 by goroutine 401:
  ??()
      -:0 +0xffffffffffffffff
  net.(*conn).Write()
      /home/harsha/sdk/go1.18.3/src/net/net.go:195 +0xa4
  net.(*TCPConn).Write()
      <autogenerated>:1 +0x64
  net.(*writerOnly).Write()
      <autogenerated>:1 +0x76
  io.copyBuffer()
      /home/harsha/sdk/go1.18.3/src/io/io.go:428 +0x2dc
  io.Copy()
      /home/harsha/sdk/go1.18.3/src/io/io.go:385 +0x8d
  net.genericReadFrom()
      /home/harsha/sdk/go1.18.3/src/net/net.go:662 +0x28
  net.(*TCPConn).readFrom()
      /home/harsha/sdk/go1.18.3/src/net/tcpsock_posix.go:54 +0xcc
  net.(*TCPConn).ReadFrom()
      /home/harsha/sdk/go1.18.3/src/net/tcpsock.go:130 +0x68
  io.copyBuffer()
      /home/harsha/sdk/go1.18.3/src/io/io.go:412 +0x1c2
  io.Copy()
      /home/harsha/sdk/go1.18.3/src/io/io.go:385 +0x86
  net/http.persistConnWriter.ReadFrom()
      /home/harsha/sdk/go1.18.3/src/net/http/transport.go:1775 +0x18
  bufio.(*Writer).ReadFrom()
      /home/harsha/sdk/go1.18.3/src/bufio/bufio.go:768 +0x28d
  io.copyBuffer()
      /home/harsha/sdk/go1.18.3/src/io/io.go:412 +0x1c2
  io.Copy()
      /home/harsha/sdk/go1.18.3/src/io/io.go:385 +0x64
  net/http.(*transferWriter).doBodyCopy()
      /home/harsha/sdk/go1.18.3/src/net/http/transfer.go:411 +0x3e
  net/http.(*transferWriter).writeBody()
      /home/harsha/sdk/go1.18.3/src/net/http/transfer.go:369 +0x799
  net/http.(*Request).write()
      /home/harsha/sdk/go1.18.3/src/net/http/request.go:698 +0x1197
  net/http.(*persistConn).writeLoop()
      /home/harsha/sdk/go1.18.3/src/net/http/transport.go:2395 +0x2e4
  net/http.(*Transport).dialConn.func6()
      /home/harsha/sdk/go1.18.3/src/net/http/transport.go:1751 +0x39

Goroutine 132 (running) created at:
  github.com/minio/minio-go/v7.Client.putObjectMultipartStreamFromReadAt()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object-streaming.go:151 +0x79d
  github.com/minio/minio-go/v7.Client.putObjectMultipartStream()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object-streaming.go:49 +0x31a
  github.com/minio/minio-go/v7.Client.putObjectCommon()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object.go:241 +0x9f7
  github.com/minio/minio-go/v7.Client.PutObject()
      /home/harsha/go/src/github.com/minio/minio-go/api-put-object.go:208 +0x384
  main.main.func1()
      /home/harsha/go/src/github.com/minio/minio-go/race-test.go:95 +0x493
  main.main.func3()
      /home/harsha/go/src/github.com/minio/minio-go/race-test.go:111 +0x47
```

You are not able to reproduce this because you are not seeing the issue with retrying that is 
triggered when one of the parallel uploads fails - since there is a Seek() call to re-upload, that
is inherently racy @bwplotka 

To avoid that Seek we have to make sure that Seeking is not allowed if you are going to
use ReaderAt without buffering.

Instead, the Seek should be converted into a ReaderAt call again to avoid offset in parallel
or add a locked "SectionReader"

I reproduced it by taking the server up and down multiple times - this triggers the race.