@bwplotka 2022-07-06
Ack. I manually injected fault to "executeMethod" to trigger the race, but unsuccessfully. 

Thanks for repro, it seems we know what we are looking for now then!